ajustado la vista para 380px

+++ src/components/CurrentDate.jsx
@@ -1,6 +1,6 @@
 export const CurrentDate = () => {
   const fecha = new Date();
-  const opciones = { weekday: 'long', month: 'short', day: '2-digit' };
+  const opciones = { weekday: 'short', month: 'short', day: '2-digit' };
   const formateada = fecha.toLocaleDateString('es-ES', opciones);
   return (
     <article className='w-full h-full bg-sl-light-olive rounded-4xl grid place-content-center py-6 px-4'>
